refactor

+++ workspace-gateway/src/main/java/com/aliware/tianchi/entity/WorkLoad.java
@@ -3,14 +3,14 @@ package com.aliware.tianchi.entity;
 public class WorkLoad {
     public int port;
 
-    public long latency;
+    public double latency;
 
-    public WorkLoad(int port, long latency) {
+    public WorkLoad(int port, double latency) {
         this.port = port;
         this.latency = latency;
     }
 
     public static int compare(WorkLoad w1, WorkLoad w2) {
-        return Long.compare(w1.latency, w2.latency);
+        return Double.compare(w1.latency, w2.latency);
     }
 }
